Example of radiosonde observations

The ExampleSonde data frame has 461 rows and 21 columns. The columns are different physical measurements made during the balloon's ascent. ExampleSondeText is an image of a text file that allows for testing the getsonde function and is an example of a format with metadata used to document a sounding.

These data are a sounding from a well known field experiment TOGA/COARE studying the larege scale dynamical relationship of the Tropical Pacific and the atmospshere. This particular sounding was launched January 17, 1993 17:12 GMT from Kavieng, New Guninea. To write out the text dataset to a sounding file just use:

data(ExampleSondeText) ; write( ExampleSondeText, file="sondedata.txt")

Format

This data frame contains the following named components:

time	Time in seconds from the ballon's release, a numeric vector.

press	Pressure (mb), a numeric vector

temp	Dry-bulb Temperature (deg. C), a numeric vector

dewpt	Dew point temperaure (deg. C), a numeric vector

rh	Relative humidity (%), a numeric vector

uwind	East-West component of the wind, a numeric vector

vwind	North-South component of the wind, a numeric vector

wspd	wind speed (m/s), a numeric vector

dir	Wind direction, a numeric vector

dz	Ascension rate (m/s), a numeric vector

lon	Longitude, a numeric vector

lat	Latitude, a numeric vector

rng	range, a numeric vector

az	azimuth (angle along horizon), a numeric vector

alt	altitude (m), a numeric vector

qp	QC flag for pressure, a numeric vector

qt	QC flag for temperature, a numeric vector

qh	QC flag for humidity, a numeric vector

qu	QC flag for U component, a numeric vector

qv	QC flag for V component, a numeric vector

quv	QC flag for ascension rate, a numeric vector

Note

Quality Control (QC) flags are generated locally at JOSS based on either automated or visual checks made.
999.0 implies the QC information is missing,
1.0 implies datum seems physically reasonable (good),
2.0 implies datum questionable on a physical basis (maybe),
3.0 implies datum seems to be in error (bad),
4.0 implies datum interpolated (estimated),
9.0 implies datum missing in original file.

See also